# openapi-my-v1.yaml
openapi: 3.0.1
info:
  title: Функции API платформы перепродажи вещей по объявлениям
  contact:
    name: Людмила Рик
    email: LRick@yandex.ru
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /upl:
    post:
      tags:
        - Объявления
      summary: Сохранить изображение
      description: Сохраняет на сервере сайта картинку, полученную от клиента.
      operationId: setImage
      requestBody:
        content:
          application/json:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
  /ads:
    get:
      tags:
        - Объявления
      summary: Список объявлений
      description: Получить список всех объявлений.
      operationId: getAllAds
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperAdsDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    post:
      tags:
        - Объявления
      summary: Новоё объявление
      description: Добавляет новоё объявление.
      operationId: addAds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdsDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdsDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
  /ads/{ad_pk}/comment:
    get:
      tags:
        - Объявления
      summary: Список комментариев к объявлению
      description: Получить список всех комментариев по заданному объявлению.
      operationId: getAllComments
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperCommentDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    post:
      tags:
        - Объявления
      summary: Новый комментарий
      description: Добавляет новый комментарий к существующему объявлению.
      operationId: addComment
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdsComment'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdsComment'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
  /ads/{ad_pk}:
    get:
      tags:
        - Объявления
      summary: Данные объявления
      description: Получить все данные по заданному объявлению.
      operationId: getAds
      parameters:
        - name: ad_pk
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FullAdsDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    delete:
      tags:
        - Объявления
      summary: Удалить объявление
      description: Удалить объявление по заданному идентификатору.
      operationId: deleteAds
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    patch:
      tags:
        - Объявления
      summary: Изменить объявление
      description: Изменить данные заданного объявления.
      operationId: updateAds
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdsDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdsDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
  /ads/{ad_pk}/comment/{id}:
    get:
      tags:
        - Объявления
      summary: Комментарий по объявлению
      description: Получить данные комментария по заданному объявлению.
      operationId: getComment
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
        - name: id
          in: path
          description: Идентификатор комментария
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdsComment'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    delete:
      tags:
        - Объявления
      summary: Удаление комментария
      description: Удаляет существующий комментарий к объявлению.
      operationId: deleteComment
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
        - name: id
          in: path
          description: Идентификатор комментария
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    patch:
      tags:
        - Объявления
      summary: Изменить комментарий
      description: Изменить содержимое комментария по заданному объявлению.
      operationId: updateComment
      parameters:
        - name: ad_pk
          in: path
          description: Ключ записи объявления
          required: true
          schema:
            type: integer
            format: int32
        - name: id
          in: path
          description: Идентификатор комментария
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdsComment'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdsComment'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
  /images/{id}/:
    get:
      tags:
        - Объявления
      summary: Получить изображение
      description: Возвращает клиенту содержимое файла картинки по её ключу.
      operationId: getImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            image/png:
              schema:
                type: array
                items:
                  type: string
                  format: byte
        '409':
          description: Conflict
        '501':
          description: Not Implemented
  /images/{id}:
    get:
      tags:
        - Объявления
      summary: Получить изображение
      description: Возвращает клиенту содержимое файла картинки по её ключу.
      operationId: getImage_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            image/png:
              schema:
                type: array
                items:
                  type: string
                  format: byte
        '409':
          description: Conflict
        '501':
          description: Not Implemented
  /ads/me:
    get:
      tags:
        - Объявления
      summary: Собственные объявления
      description: Получить список всех собственных объявлений.
      operationId: getAdsMe
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperAdsDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
components:
  schemas:
...
    CreateAdsDto:
      type: object
      properties:
        description:
          type: string
          description: Текст описания товара по объявлению
        image:
          type: string
          description: Ссылка для скачивания содержимого картинки
        pk:
          type: integer
          description: Первичный ключ записи объявления
          format: int32
        price:
          type: integer
          description: Объявленная цена за товар
          format: int32
        title:
          type: string
          description: Заголовок объявления
      description: Данные нового объявления
    AdsDto:
      type: object
      properties:
        pk:
          type: integer
          description: Первичный ключ записи объявления
          format: int32
        author:
          type: integer
          description: Идентификатор автора объявления
          format: int32
        price:
          type: integer
          description: Объявленная цена за товар
          format: int32
        title:
          type: string
          description: Заголовок объявления
        image:
          type: string
          description: Ссылка для скачивания содержимого картинки
      description: Данные объявления
    AdsComment:
      type: object
      properties:
        pk:
          type: integer
          description: Первичный ключ записи объявления
          format: int32
        author:
          type: integer
          description: Идентификатор автора комментария
          format: int32
        createdAt:
          type: string
          description: Дата и время создания комментария
          format: date-time
        text:
          type: string
          description: Текст комментария
      description: Данные комментария
...
    ResponseWrapperAdsDto:
      type: object
      properties:
        count:
          type: integer
          description: Количество записей
          format: int32
        results:
          type: array
          description: Список данных пользователя
          items:
            $ref: '#/components/schemas/AdsDto'
      description: Оболочка ответа о данных объявления
    FullAdsDto:
      type: object
      properties:
        authorFirstName:
          type: string
          description: Имя автора объявления
        authorLastName:
          type: string
          description: Фамилия автора объявления
        description:
          type: string
          description: Текст описания товара по объявлению
        email:
          type: string
          description: Адрес электронной почты автора объявления
        image:
          type: string
          description: Ссылка для скачивания содержимого картинки
        phone:
          type: string
          description: Контактный телефон автора объявления
        pk:
          type: integer
          description: Первичный ключ записи объявления
          format: int32
        price:
          type: integer
          description: Объявленная цена за товар
          format: int32
        title:
          type: string
          description: Заголовок объявления
      description: Полная информация об объявлении и его авторе
    ResponseWrapperCommentDto:
      type: object
      properties:
        count:
          type: integer
          description: Количество записей
          format: int32
        results:
          type: array
          description: Список данных пользователя
          items:
            $ref: '#/components/schemas/AdsComment'
      description: Оболочка ответа по комментариям к объявлению